feat: add redis

+++ src/application/services/event/commands/delete-event.ts
@@ -1,5 +1,7 @@
-import { NotFoundException } from '@domain/common';
+import { Roles } from '@common/config/roles';
+import { NotFoundException, NotRightsException } from '@domain/common';
 import { IEventRepository } from '@domain/event';
+import { IUserRepository } from '@domain/user';
 import { Result } from 'true-myth';
 
 import { UUID } from 'crypto';
@@ -7,18 +9,40 @@ import { UUID } from 'crypto';
 import { safeAsync } from '../../common';
 
 export class DeleteEvent {
-    constructor(readonly eventId: UUID) {}
+    constructor(
+        readonly role: Roles,
+        readonly userId: UUID,
+        readonly eventId: UUID,
+    ) {}
 }
 
 export class DeleteEventHandler {
-    constructor(private readonly eventRepo: IEventRepository) {}
+    constructor(
+        private readonly userRepo: IUserRepository,
+        private readonly eventRepo: IEventRepository,
+    ) {}
 
     execute(command: DeleteEvent): Promise<Result<boolean, Error>> {
         return safeAsync(async () => {
+            const requestUser = await this.userRepo.findById(command.userId);
+            if (!requestUser) throw new NotFoundException();
+
+            if (requestUser.isBlocked) {
+                throw new NotRightsException();
+            }
+
             const event = await this.eventRepo.findById(command.eventId);
             if (!event) throw new NotFoundException();
 
+            const isAdmin = command.role === Roles.ADMIN;
+            const isOwner = event.canEditedBy(requestUser.id);
+
+            if (!isAdmin && !isOwner) {
+                throw new NotRightsException();
+            }
+
             await this.eventRepo.delete(event.id);
+
             return true;
         });
     }
