+++ src/application/services/user/commands/delete-user.ts
@@ -1,4 +1,5 @@
-import { NotFoundException } from '@domain/common';
+import { Roles } from '@common/config/roles';
+import { NotFoundException, NotRightsException } from '@domain/common';
 import { IUserRepository } from '@domain/user';
 import { Result } from 'true-myth';
 
@@ -7,7 +8,10 @@ import { UUID } from 'crypto';
 import { safeAsync } from '../../common';
 
 export class DeleteUser {
-    constructor(readonly userId: UUID) {}
+    constructor(
+        readonly role: Roles,
+        readonly userId: UUID,
+    ) {}
 }
 
 export class DeleteUserHandler {
@@ -17,6 +21,9 @@ export class DeleteUserHandler {
         return safeAsync(async () => {
             const user = await this.userRepo.findById(command.userId);
             if (!user) throw new NotFoundException();
+            if (command.role !== Roles.ADMIN && !user.canEditedBy(command.userId)) {
+                throw new NotRightsException();
+            }
 
             await this.userRepo.delete(user.id);
